refactor(client): Move tinker_utils to common and eliminate sitecustomize magic
- Moves examples/autoresearch/tinker_utils.py to examples/common/tinker_utils.py
  and registers common in examples/pyproject.toml so utilities can be imported
  from anywhere inside examples/ without coupling to autoresearch.
- Deletes examples/sitecustomize.py to eliminate implicit runtime monkeypatching.
- Updates math_rl and text_sql recipes to explicitly invoke
  patch_tinker_default_headers() on startup.
- Updates scripts/run_training_e2e.py to invoke the local recipes.math_rl.train
  wrapper for non-Gemma math RL runs, ensuring custom headers are propagated.

TAG=agy
CONV=e300f754-575f-434b-9b5b-4d3923ccf8c8
