+++ examples/pyproject.toml
@@ -50,10 +50,11 @@ override-dependencies = [
 
 [tool.setuptools.package-dir]
 "" = "autoresearch"
+"common" = "common"
 
 [tool.setuptools]
-py-modules = ["event_log", "run_attempt", "telemetry", "tinker_utils"]
+py-modules = ["event_log", "run_attempt", "telemetry"]
 
 [tool.setuptools.packages.find]
-where = ["autoresearch"]
-include = ["recipes*", "ui*"]
+where = ["autoresearch", "."]
+include = ["recipes*", "ui*", "common*"]
