+++ scripts/run_training_e2e.py
@@ -529,8 +529,8 @@ def _math_rl_train_module_and_renderer(base_model: str) -> tuple[str, str]:
   if "gemma" in base_model.lower():
     return "recipes.math_rl.train_gemma", "gemma4"
   if "Instruct" in base_model or "Qwen2.5" in base_model:
-    return "tinker_cookbook.recipes.math_rl.train", "qwen3_instruct"
-  return "tinker_cookbook.recipes.math_rl.train", "qwen3"
+    return "recipes.math_rl.train", "qwen3_instruct"
+  return "recipes.math_rl.train", "qwen3"
 
 
 def run_gsm8k_rl(config: RunConfig, base_url: str, watch: list[ManagedProcess]) -> None:
